Initial commit
'rails new railsgirls'

Note: There are a few differences in my initial commit then you will
see in yours when following the tutorial. That is because I use
slightly different tools. For now you can ignore the differences
because the tools accomplish the same thing. If you want to know
more about the differences, Google for rbenv and Bundler.