Step 5: Create root path
I did this step differently because using a redirect in routes
is arguably a bad practice. When we say redirect, we are sending
a message to the browser to go to a different url. Instead, what
we want to do is render a specific page. This way there is only
one request to the server instead of two.